chore: reorganizing components
