+++ src/components/GlobalStyle/index.js
