--- src/components/animation/index.js
@@ -1,7 +1,0 @@
-import growToRight from '../animation/GrowToRight';
-import decreaseToLeft from '../animation/DecreaseToLeft';
-
-export {
-  growToRight,
-  decreaseToLeft
-}
