--- src/components/header/index.js
@@ -1,7 +1,0 @@
-import TopBar from './TopBar';
-import TextFieldHeader from './TextFieldHeader';
-
-export {
-  TopBar,
-  TextFieldHeader
-}
