+++ src/components/index.js
@@ -1,0 +1,13 @@
+import GlobalStyle from './GlobalStyle';
+import Header from './Header';
+import TextFieldHeader from './TextFieldHeader';
+import TopBar from './TopBar';
+
+
+export {
+  GlobalStyle,
+  Header,
+  
+  TextFieldHeader,
+  TopBar
+}
